+++ test/Node/NpmLockSpec.hs
@@ -1,13 +1,36 @@
+{-# LANGUAGE TemplateHaskell #-}
+
 module Node.NpmLockSpec (
   spec,
 ) where
 
 import Data.Map.Strict qualified as Map
 import Data.Set qualified as Set
-import DepTypes
-import GraphUtil
-import Strategy.Node.Npm.PackageLock
-import Test.Hspec
+import DepTypes (
+  DepEnvironment (EnvDevelopment, EnvProduction),
+  DepType (NodeJSType),
+  Dependency (..),
+  VerConstraint (CEq),
+ )
+import Effect.ReadFS (readContentsJson)
+import GraphUtil (expectDeps, expectDirect, expectEdges)
+import Path (mkRelDir, mkRelFile, (</>))
+import Path.IO (getCurrentDir)
+import Strategy.Node.Npm.PackageLock (
+  NpmDep (
+    NpmDep,
+    depDependencies,
+    depDev,
+    depRequires,
+    depResolved,
+    depVersion
+  ),
+  NpmPackageJson (..),
+  NpmResolved (NpmResolved, unNpmResolved),
+  buildGraph,
+ )
+import Test.Effect (it', shouldBe')
+import Test.Hspec (Spec, describe, it, runIO)
 
 mockInput :: NpmPackageJson
 mockInput =
@@ -19,7 +42,7 @@ mockInput =
             , NpmDep
                 { depVersion = "1.0.0"
                 , depDev = False
-                , depResolved = Just "https://example.com/one.tgz"
+                , depResolved = NpmResolved $ Just "https://example.com/one.tgz"
                 , depRequires = Map.fromList [("packageTwo", "2.0.0"), ("packageSeven", "7.0.0")]
                 , depDependencies =
                     Map.fromList
@@ -28,7 +51,7 @@ mockInput =
                         , NpmDep
                             { depVersion = "2.0.0"
                             , depDev = True
-                            , depResolved = Just "https://example.com/two.tgz"
+                            , depResolved = NpmResolved $ Just "https://example.com/two.tgz"
                             , depRequires = Map.fromList [("packageThree", "3.0.0")]
                             , depDependencies = mempty
                             }
@@ -41,7 +64,7 @@ mockInput =
             , NpmDep
                 { depVersion = "3.0.0"
                 , depDev = True
-                , depResolved = Just "https://example.com/three.tgz"
+                , depResolved = NpmResolved $ Just "https://example.com/three.tgz"
                 , depRequires = Map.fromList [("packageOne", "1.0.0")]
                 , depDependencies = mempty
                 }
@@ -51,7 +74,7 @@ mockInput =
             , NpmDep
                 { depVersion = "7.0.0"
                 , depDev = False
-                , depResolved = Just "https://example.com/seven.tgz"
+                , depResolved = NpmResolved $ Just "https://example.com/seven.tgz"
                 , depRequires = mempty
                 , depDependencies = mempty
                 }
@@ -61,7 +84,7 @@ mockInput =
             , NpmDep
                 { depVersion = "file:abc/def"
                 , depDev = False
-                , depResolved = Nothing
+                , depResolved = NpmResolved Nothing
                 , depRequires = mempty
                 , depDependencies = mempty
                 }
@@ -71,7 +94,7 @@ mockInput =
             , NpmDep
                 { depVersion = "5.0.0"
                 , depDev = True
-                , depResolved = Just "https://example.com/five.tgz"
+                , depResolved = NpmResolved $ Just "https://example.com/five.tgz"
                 , depRequires = Map.fromList [("packageSix", "6.0.0")]
                 , depDependencies = mempty
                 }
@@ -81,7 +104,7 @@ mockInput =
             , NpmDep
                 { depVersion = "6.0.0"
                 , depDev = True
-                , depResolved = Just "https://example.com/six.tgz"
+                , depResolved = NpmResolved $ Just "https://example.com/six.tgz"
                 , depRequires = mempty
                 , depDependencies = mempty
                 }
@@ -170,3 +193,24 @@ spec = do
         , (packageFive, packageSix)
         ]
         graph
+  describe "parsing package-json.lock" $ do
+    currentDir <- runIO getCurrentDir
+    let testDir = currentDir </> $(mkRelDir "test/Node/testdata")
+
+    it' "Should ignore \"resolved\": <bool> in package-lock.json" $ do
+      let packageLock = testDir </> $(mkRelFile "boolean-resolved-package-lock.json")
+      NpmPackageJson{packageDependencies = packageDependencies} <- readContentsJson packageLock
+      let foo = unNpmResolved . depResolved =<< Map.lookup "foo" packageDependencies
+      foo `shouldBe'` Nothing
+
+    it' "Should parse \"resolved\": <string> in package-lock.json" $ do
+      let packageLock = testDir </> $(mkRelFile "string-resolved-package-lock.json")
+      NpmPackageJson{packageDependencies = packageDependencies} <- readContentsJson packageLock
+      let foo = unNpmResolved . depResolved =<< Map.lookup "foo" packageDependencies
+      foo `shouldBe'` Just "https://bar.npmjs.org/foo/-/foo-1.0.0.tgz"
+
+    it' "Should parse dependency with no \"resolved\" key in package-lock.json" $ do
+      let packageLock = testDir </> $(mkRelFile "absent-resolved-package-lock.json")
+      NpmPackageJson{packageDependencies = packageDependencies} <- readContentsJson packageLock
+      let foo = unNpmResolved . depResolved =<< Map.lookup "foo" packageDependencies
+      foo `shouldBe'` Nothing